+++ public/modules/organizations/controllers/org-controller.js
@@ -1,8 +1,8 @@
 angular.module('bitvagas.org.controllers',[])
 .controller('OrgController', OrgController);
 
-OrgController.$inject = ['$scope', '$state', 'OrgService', 'ErrorHandling'];
-function OrgController($scope, $state, OrgService, ErrorHandling){
+OrgController.$inject = ['$scope', '$state', 'OrgService', 'SweetAlert'];
+function OrgController($scope, $state, OrgService, SweetAlert){
 
     if($state.params.OrgID !== undefined) {
         OrgService.findById($state.params.OrgID)
@@ -16,7 +16,6 @@ function OrgController($scope, $state, OrgService, ErrorHandling){
             GoBack();
         }, function(err){
             console.log(err);
-            $scope.errors = ErrorHandling.getErrors(err.data);
         });
     };
 
@@ -25,18 +24,39 @@ function OrgController($scope, $state, OrgService, ErrorHandling){
             GoBack();
         }, function(err){
             console.log(err);
-            $scope.errors = ErrorHandling.getErrors(err.data);
         });
     };
 
     $scope.delete = function(id){
-        //TODO: Message confirmation
-        OrgService.delete(id).then(function(data){
-            GoBack();
-        }, function(err){
-            console.log(err);
-            $scope.errors = ErrorHandling.getErrors(err.data);
-        });
+
+        SweetAlert.swal({
+            title: "Delete Organization"
+          , text: "Deseja deletar essa organização"
+          , type: "warning"
+          , showCancelButton: true
+          , confirmButtonColor: "#DD6B55",confirmButtonText: "Sim, deletar!"
+          , cancelButtonText: "No"
+          , closeOnConfirm: false
+          , closeOnCancel: false }
+          , function(isConfirm){
+                if (isConfirm) {
+                    SweetAlert.swal({ title: "Deleted!"
+                                    , text: "Organização deletada."
+                                    , type: "success"
+                                    , confirmButtonColor: "#29B5DF"
+                                    });
+                    OrgService.delete(id).then(function(data){
+                        GoBack();
+                    }, function(err){
+                        console.log(err);
+                    });
+                } else {
+                    SweetAlert.swal({ title: "Cancelled"
+                                    , type: "error"
+                                    , confirmButtonColor: "#29B5DF"
+                                    });
+                }
+            });
     };
 
     function GoBack(){
